* fix permission level from 400 to 600 for the SSH pem file (#9)

* storing ssh private key in secrets manager

* changing carts microservice to use ddb instead of ddb local (#8)

* create local ssh config for mac only if the path doesn't match cloud9

* Enable ack (#10)

* Add ack support in terraform


---------


* Staging vs prod (#11)

* staging vs. prod deployments

---------


* adding kms key admins to the TF config

* Refactor with Tenants (#13)

* refactor for tenants

---------


* add support for additional roles for KMS and aws-auth configmap

* terraform init comes before workspace command

* missing env var for GITOPS REPOS
fix CodeCommit TF apply comand for Cloud9

* add missing namespaces to resources (#14)


* refactor addons folder (#15)

* Riv23 refactor platform folder (#16)

* refactor platform folder

* changing platform repo path for the appset secret labekl

* update `setup-git.sh` script

* Carts - starting with self-managed and modifying the structure to simplify switch to managed (#17)

* carts - starting with self-managed and modifying the structure to simplify switch to managed

* using placeholder for account id

* changing workload appsets to allow auto sync to delete resources

* pin eks module to 19.19.0

* update yaml files dirs (#19)


* Adding separate kustomization.yaml for managed ddb deployment (#18)

* carts - starting with self-managed and modifying the structure to simplify switch to managed

* using placeholder for account id

* changing workload appsets to allow auto sync to delete resources

* adding separate kustomization.yaml for managed ddb deployment

---------


* create a solution file for addons in spoke cluster instead of `sed`-ing the file

* add name for argo AppProjects (#20)

* remove unused resource in In terraform/codecommit/main.tf

* adding limit ranges and resource quotas to all app namespaces

* delete bkp folders (#21)


* Carrlos updates 11 17 (#22)

* test argo


* remove embedded argoproject chart


* update both projects


* nothing to exclude in the workshop


* update load balancer version for dev


* basepath


* add the infra mng kustomize for carts


---------


* changing minReplicas and maxReplicas for ui to better demonstrate prod config changes in the workshop lab (#23)

* updating instance type to c5.large to prevent orders service crashloopbackoff

* changing argocd endpoint from https to http (#25)

---------

+++ terraform/spokes/destroy.sh
@@ -6,17 +6,6 @@ SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 ROOTDIR="$(cd ${SCRIPTDIR}/../..; pwd )"
 [[ -n "${DEBUG:-}" ]] && set -x
 
-
-pushd () {
-    command pushd "$@" > /dev/null
-}
-
-popd () {
-    command popd "$@" > /dev/null
-}
-
-pushd ${SCRIPTDIR}
-
 if [[ $# -eq 0 ]] ; then
     echo "No arguments supplied"
     echo "Usage: destroy.sh <environment>"
@@ -26,19 +15,20 @@ fi
 env=$1
 echo "Destroying $env ..."
 
-set -x
-
-terraform workspace select $env
+terraform -chdir=$SCRIPTDIR workspace select $env
 # Delete the Ingress/SVC before removing the addons
 TMPFILE=$(mktemp)
-terraform output -raw configure_kubectl > "$TMPFILE"
-source "$TMPFILE"
-kubectl delete svc --all -n ui
+terraform -chdir=$SCRIPTDIR output -raw configure_kubectl > "$TMPFILE"
+# check if TMPFILE contains the string "No outputs found"
+if [[ ! $(cat $TMPFILE) == *"No outputs found"* ]]; then
+  echo "No outputs found, skipping kubectl delete"
+  source "$TMPFILE"
+  kubectl delete svc --all -n ui
+  kubectl delete -A tables.dynamodb.services.k8s.aws --all
+fi
 
-terraform destroy -target="module.gitops_bridge_bootstrap" -auto-approve -var-file="workspaces/${env}.tfvars"
-terraform destroy -target="module.eks_blueprints_addons" -auto-approve -var-file="workspaces/${env}.tfvars"
-terraform destroy -target="module.eks" -auto-approve -var-file="workspaces/${env}.tfvars"
-terraform destroy -target="module.vpc" -auto-approve -var-file="workspaces/${env}.tfvars"
-terraform destroy -auto-approve -var-file="workspaces/${env}.tfvars"
-
-popd
+terraform -chdir=$SCRIPTDIR destroy -target="module.gitops_bridge_bootstrap_hub" -auto-approve -var-file="workspaces/${env}.tfvars"
+terraform -chdir=$SCRIPTDIR destroy -target="module.eks_blueprints_addons" -auto-approve -var-file="workspaces/${env}.tfvars"
+terraform -chdir=$SCRIPTDIR destroy -target="module.eks" -auto-approve -var-file="workspaces/${env}.tfvars"
+terraform -chdir=$SCRIPTDIR destroy -target="module.vpc" -auto-approve -var-file="workspaces/${env}.tfvars"
+terraform -chdir=$SCRIPTDIR destroy -auto-approve -var-file="workspaces/${env}.tfvars"
